add 3 and 4 module

+++ 170877_Harsh_DR_Module1_T006/src/170877_harsh_DR_Module1_T006_1-22.c
@@ -21,7 +21,7 @@ int16_t c;
 int8_t fold;
 char line[maxline];
 int8_t col = 0;
-int8_t i = 0;
+int i = 0;
 
 printf("After how much character you want to fold : ");
 scanf("%hhd",&fold);
